Working NVVM/CUDA support on Tegra K1:
- use platform independent variables of fixed size in codegen/runtime

+++ runtime/cpu/cpu_runtime.cpp
@@ -5,7 +5,7 @@
 
 // helper functions
 void thorin_init() { }
-void *thorin_malloc(size_t size) {
+void *thorin_malloc(uint32_t size) {
     void *mem;
     posix_memalign(&mem, 64, size);
     std::cerr << " * malloc(" << size << ") -> " << mem << std::endl;
